requirements-trainer.txt
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
google-generativeai>=0.3.2
qdrant-client>=1.7.0
pandas>=2.1.4
faker>=20.1.0
python-dotenv>=1.0.0
requests>=2.31.0